abstracted reporting class
